Added ParkourWarrior
- Added ParkourWarrior
- Changed points type to double
- Updated mysql dependencies
- Added placeholders

+++ src/main/java/ink/ziip/championshipscore/api/game/area/team/BaseTeamArea.java
@@ -51,7 +51,7 @@ public abstract class BaseTeamArea extends BaseArea {
 
     @Override
     public void addPlayerPointsToDatabase() {
-        for (Map.Entry<UUID, Integer> playerPointEntry : playerPoints.entrySet()) {
+        for (Map.Entry<UUID, Double> playerPointEntry : playerPoints.entrySet()) {
             if (playerPointEntry.getValue() != 0) {
                 ChampionshipTeam championshipTeam = plugin.getTeamManager().getTeamByPlayer(playerPointEntry.getKey());
                 if (championshipTeam != null) {
